Assignment 7: NumPy Pandas Matplotlib

+++ week-5/task-2.py
@@ -1,14 +1,11 @@
 import json
 
-# Read original JSON
 with open("students.json", "r", encoding="utf-8") as file:
     students = json.load(file)
 
-# Calculate average grade
 for student in students:
     grades = student["grades"]
-    student["average_grade"] = sum(grades) / len(grades)
+    student["average_grade"] = int(sum(grades) / len(grades))
 
-# Write to a new JSON file
 with open("students1.json", "w", encoding="utf-8") as file:
     json.dump(students, file, indent=4)
